# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Directorio de empresas - CCGAMT",
        "description": "Backend y API Acceso a datos.",
        "contact": {
            "email": "brayan-diaz3@upc.edu.co"
        },
        "license": {
            "name": "@Universidad Piloto de Colombia SAM - Todos los Derechos Reservados"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/lecciones": {
            "get": {
                "tags": [
                    "Lecciones"
                ],
                "summary": "Listar todas las lecciones",
                "operationId": "3098182027fbb7166466d536ae38482c",
                "responses": {
                    "200": {
                        "description": "Lista de lecciones"
                    }
                }
            },
            "post": {
                "tags": [
                    "Lecciones"
                ],
                "summary": "Crear una nueva lección",
                "operationId": "54b94bbfc77ce9c3934f30f6929c2efc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "id_nivel"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Lección 1"
                                    },
                                    "contenido": {
                                        "type": "string",
                                        "example": "Contenido de la lección"
                                    },
                                    "id_nivel": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lección creada"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/lecciones/{id}": {
            "get": {
                "tags": [
                    "Lecciones"
                ],
                "summary": "Obtener una lección específica",
                "operationId": "24fcd9fbd2ae1deffb52a5812e730d2f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la lección",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lección encontrada"
                    },
                    "404": {
                        "description": "Lección no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Lecciones"
                ],
                "summary": "Actualizar una lección existente",
                "operationId": "35f3d816bb9443667e1af67e54dc43f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la lección",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Lección actualizada"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Contenido actualizado"
                                    },
                                    "curso_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lección actualizada"
                    },
                    "404": {
                        "description": "Lección no encontrada"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Lecciones"
                ],
                "summary": "Eliminar una lección",
                "operationId": "6d01d40d6e1c185978b4457d0309ca48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la lección",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lección eliminada correctamente"
                    },
                    "404": {
                        "description": "Lección no encontrada"
                    }
                }
            }
        },
        "/api/niveles": {
            "get": {
                "tags": [
                    "Niveles"
                ],
                "summary": "Listar todos los niveles",
                "operationId": "6e707cb14cb81a42caf114e02857411f",
                "responses": {
                    "200": {
                        "description": "Lista de niveles"
                    }
                }
            },
            "post": {
                "tags": [
                    "Niveles"
                ],
                "summary": "Crear un nuevo nivel",
                "operationId": "19a09347d504945147aedc3774c49003",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Nivel Básico"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Primer nivel de dificultad"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nivel creado exitosamente"
                    },
                    "400": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/niveles/{id}": {
            "get": {
                "tags": [
                    "Niveles"
                ],
                "summary": "Obtener un nivel específico",
                "operationId": "b32dc128c14f03c334577ac9bc7b7f1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del nivel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nivel encontrado"
                    },
                    "404": {
                        "description": "Nivel no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Niveles"
                ],
                "summary": "Actualizar un nivel existente",
                "operationId": "701bf5f1f82b1a6f32a0f3e41c6145a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del nivel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Nivel Intermedio"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Segundo nivel de dificultad"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nivel actualizado exitosamente"
                    },
                    "400": {
                        "description": "Errores de validación"
                    },
                    "404": {
                        "description": "Nivel no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Niveles"
                ],
                "summary": "Eliminar un nivel",
                "operationId": "d798a7eb8592900e6657fdb3cc958bd8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del nivel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nivel eliminado"
                    },
                    "404": {
                        "description": "Nivel no encontrado"
                    }
                }
            }
        },
...
    },
...
}